Move Array to its own module (#61)
* Move Array to its own module
