+++ docs/apidocs/array.rst
@@ -1,0 +1,6 @@
+.. _qiskit_dynamics-array:
+
+.. automodule:: qiskit_dynamics.array
+   :no-members:
+   :no-inherited-members:
+   :no-special-members:
