+++ docs/tutorials/optimizing_pulse_sequence.rst
@@ -111,7 +111,7 @@ example of one.
 .. jupyter-execute::
 
     from qiskit_dynamics import DiscreteSignal
-    from qiskit_dynamics.dispatch import Array
+    from qiskit_dynamics.array import Array
     from qiskit_dynamics.signals import Convolution
 
     # define convolution filter
