+++ qiskit_dynamics/array/patch_qi.py
