+++ qiskit_dynamics/models/operator_collections.py
@@ -21,7 +21,7 @@ from scipy.sparse.csr import csr_matrix
 
 from qiskit import QiskitError
 from qiskit.quantum_info.operators.operator import Operator
-from qiskit_dynamics.dispatch import Array
+from qiskit_dynamics.array import Array
 from qiskit_dynamics.type_utils import to_array, to_csr, vec_commutator, vec_dissipator
