+++ qiskit_dynamics/models/rotating_frame.py
@@ -22,7 +22,7 @@ from scipy.sparse import issparse, csr_matrix
 from qiskit import QiskitError
 from qiskit.quantum_info.operators import Operator
 from qiskit.quantum_info.operators.predicates import is_hermitian_matrix
-from qiskit_dynamics.dispatch import Array
+from qiskit_dynamics.array import Array
 from qiskit_dynamics.type_utils import to_array, to_numeric_matrix_type
