+++ qiskit_dynamics/solvers/jax_odeint.py
@@ -21,7 +21,8 @@ from typing import Callable, Optional, Union, Tuple, List
 import numpy as np
 from scipy.integrate._ivp.ivp import OdeResult
 
-from qiskit_dynamics.dispatch import requires_backend, Array, wrap
+from qiskit_dynamics.dispatch import requires_backend
+from qiskit_dynamics.array import Array, wrap
 
 from .solver_utils import merge_t_args, trim_t_results
