+++ qiskit_dynamics/type_utils.py
@@ -27,7 +27,7 @@ from scipy.sparse.csr import csr_matrix
 
 from qiskit.quantum_info.operators import Operator
 from qiskit_dynamics import dispatch
-from qiskit_dynamics.dispatch import Array
+from qiskit_dynamics.array import Array
 
 
 class StateTypeConverter:
