+++ test/dynamics/models/test_lindblad_model.py
@@ -23,7 +23,7 @@ from qiskit import QiskitError
 from qiskit.quantum_info.operators import Operator
 from qiskit_dynamics.models import LindbladModel
 from qiskit_dynamics.signals import Signal, SignalList
-from qiskit_dynamics.dispatch import Array
+from qiskit_dynamics.array import Array
 from qiskit_dynamics import dispatch
 from ..common import QiskitDynamicsTestCase, TestJaxBase
