+++ test/dynamics/test_type_utils.py
@@ -18,7 +18,7 @@ import numpy as np
 from scipy.sparse.csr import csr_matrix
 
 from qiskit.quantum_info.operators.operator import Operator
-from qiskit_dynamics.dispatch import Array
+from qiskit_dynamics.array import Array
 from qiskit_dynamics.type_utils import (
     convert_state,
     type_spec_from_instance,
